Added dummy route test. Checking if CIrcle CI gives an error
