+++ package.json
@@ -3,6 +3,9 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
+  "jest": {
+    "verbose": true
+  },
   "scripts": {
     "start": "nodemon server/index.js",
     "test": "jest",
@@ -31,6 +34,7 @@
     "jest-cli": "^27.0.6",
     "mongoose": "^5.13.2",
     "nodemon": "^2.0.9",
-    "pg": "^8.6.0"
+    "pg": "^8.6.0",
+    "supertest": "^6.1.4"
   }
 }
